+++ packages/components/src/components/LightBox/LightBox.module.d.scss.ts
@@ -1,5 +1,7 @@
 declare const classNames: {
   readonly lightBox: "lightBox";
+  readonly "flow--overlay--viewport": "flow--overlay--viewport";
+  readonly "react-aria-Modal": "react-aria-Modal";
   readonly content: "content";
   readonly fitScreen: "fitScreen";
   readonly actions: "actions";